Refactor people query into separate file.
